+++ lib/pronto/formatter/bitbucket_formatter.rb
@@ -9,7 +9,7 @@ module Pronto
         'BitBucket'
       end
 
-      def line_number(message)
+      def line_number(message, _)
         message.line.new_lineno
       end
     end