+++ lib/pronto/formatter/github_pull_request_formatter.rb
@@ -8,6 +8,11 @@ module Pronto
       def pretty_name
         'GitHub'
       end
+
+      def line_number(message, patches)
+        line = patches.find_line(message.full_path, message.line.new_lineno)
+        line.position
+      end
     end
   end
 end